Functions can be args, added builtin procs
Functions can be passed as arguments.
To acomplish this a new symbol type was created, the ambiguous symbol.
This symbol can either be a constant or a procedure and is determined at
runtime. When the semantic analyzer reaches a procedure assignment, the
formal paramaters map to ambiguous symbols, so that in further
evaluation they can be treated as procedures or constants. Since the
identity of arguments which are used as functions are determined at
runetime during procedure calls, semantic analysis could no longer occur
before interpretation. The semantic analyzer was therefore moved inside
the interpreter, and they have references to each other. Semantic
analysis is now performed as the tree is interpreted so that in the case
when a formal paramter is treated as a procedure, the actual paramter
can be determined through a somewhat complicated processes of looking
through the call-stack and scoped symbol table that I still have trouble
understanding.

The 'if' function and the '=' function were implemented, the purpose of
which was to implement the factorial function. The main function has an
example of this, along with a somewhat confusing example at the top that
tests scoping. At the bottom there are examples of procedures being
used as actual parameters.

The next step is to focus on refactoring the code and adding better
documentation. After that I would like to enhance what an ID token means
which should get rid of the special token types for single character
operators. I would also like to consolidate where builtin functions are
defined into one file so places that refer to builtin functions can
refer back there instead of having files like the interpreter have long
functions just to deal with builtin functions. I would also like to
implement checks for builtin functions so that I can define how many
parameters they should take in. Then I want to enhance the Number data
type so that there are precise and imprecise numbers.