Fixed some bugs, chained __main__ code
The code in the main file is now some code that I wrote in university,
modified a bit to fit the interpreter which doesn't have all the
required features to interpret the code. The changes are noted in the
__main__ file, commented out at the top of the code.

Pretty excited that the interpreter is able to understand the code.