Restructured lexing process
The end goal of this change is to be able to process lists that use the
notation starting with an apostrophe. Since using an apostrophe is for
the most part reserved for symbols, the initial design of the lexer was
very limited. The new structure should make it easier to extend what is
accepted by the lexer. Nothing has yet been added, just some minor
restructuring.