Changed how symbols are interpreted
The lexer now only returns a quote whenever
one is reached instead of the entire symbol.
The building of the symbol or list or data
is now done at the parsing level. Symbols
also take in the name without the quote,
and the string function will take care of
including the quote. This is also done
because in the future when printing lists
with symbols, the leading quote should not
be included in the symbol.

There is still a lot to be done in the
parser for building lists with the list
shorthand.